+++ examples/scripts/3d_examples/3d_basic_spm_vs_lumped.py
@@ -6,7 +6,7 @@ import pybamm
 # Load models
 models = {
     "Lumped": pybamm.lithium_ion.SPM(options={"thermal": "lumped"}),
-    "3D": pybamm.lithium_ion.BasicSPM_with_3DThermal(
+    "3D": pybamm.lithium_ion.Basic3DThermalSPM(
         options={"cell geometry": "pouch", "dimensionality": 3}
     ),
 }