+++ src/pybamm/models/full_battery_models/lithium_ion/__init__.py
@@ -19,7 +19,7 @@ from .dfn import DFN
 from .newman_tobias import NewmanTobias
 from .basic_dfn import BasicDFN
 from .basic_spm import BasicSPM
-from .basic_spm_with_3d_thermal import BasicSPM_with_3DThermal
+from .basic_spm_with_3d_thermal import Basic3DThermalSPM
 from .basic_dfn_half_cell import BasicDFNHalfCell
 from .basic_dfn_composite import BasicDFNComposite
 from .Yang2017 import Yang2017